Fix stats search without time period and move business logic from router to service layer
